Correct git user name/email config

+++ dev_setup.sh
@@ -1,12 +1,15 @@
 #!/bin/sh
 
+# Requires the PATH variable to be updated 
+# Can be done by modifying $HOME/.bash_profile
+# PATH=$PATH:$HOME/bin:/usr/sbin:/sbin:$HOME/git/bin
+
 if [[($1 = "") || ($2 = "")]]; then
     echo "Usage: dev_setup.sh <git user.name> <git user.email>"
     exit 1
 fi
 
 # Install git
-
 git_version=1.8.1.4
 cd $HOME
 wget http://git-core.googlecode.com/files/git-$git_version.tar.gz 
@@ -15,14 +18,13 @@ cd git-$git_version
 ./configure --prefix=$HOME/git 
 make 
 make install 
-export PATH=$PATH:$HOME/git/bin
 cd $HOME
 rm -rf git-$git_version
 rm git-$git_version.tar.gz
 
 # Configure git user name and email
-git config --global user.name=$1
-git config --global user.email=$2
+git config --global user.name "$1"
+git config --global user.email "$2"
 
 # Setup the public key
 ssh-keygen
@@ -33,9 +35,15 @@ echo
 read
 
 # Clone the MicroCloud repo
-cd $HOmE
+cd $HOME
 git clone git@github.com:mschuma/MicroCloud.git
-git status
 
-# Copy .vimrc
-cp .vimrc $HOME/.vimrc
+git_repo_name="MicroCloud"
+if [ -d $git_repo_name ]; then
+    cd MicroCloud
+
+    if [ -f ".vimrc" ]; then    
+        # Copy .vimrc
+        cp .vimrc $HOME/.vimrc
+    fi
+fi
